requirements.txt
aiozipstream==0.4
anyio==3.6.1
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
elastic-transport==8.1.2
elasticsearch==7.10.0
fastapi==0.78.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
pydantic==1.9.1
python-multipart==0.0.5
requests==2.28.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
stream-zip==0.0.47
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
